scraper/requirements.txt
scrapy>=2.11.0
beautifulsoup4>=4.12.2
requests>=2.31.0
pydantic>=2.9.0  # Python 3.13 compatible
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9  # PostgreSQL driver for backfill script
